scraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
lxml==5.1.0
selenium==4.16.0
icalendar==5.0.11
